+++ CHANGELOG.md
@@ -18,6 +18,8 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 - Wikidata IDs are now also loaded into recordedByID and identifiedByID [#3989]
 - Sort order of descendant inventory
 - Removed net-pop gem workaround for Ruby 3.3.3
+- Facet geographic area: Spatial mode by default
+- Facet nomenclature rank: Remove selected ranks from select input
 
 ### Fixed
 
@@ -27,7 +29,6 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 - Quickly clicking save before load-in on Comprehensive can detach CollectingEvent from CollectionObject
 - Filter collecting events: data attribute table view is empty [#4013]
 - Field synchronize: URI Too Large error when user pass a long query [#4017]
-
 
 [#4019]: https://github.com/SpeciesFileGroup/taxonworks/issues/4019
 [#3989]: https://github.com/SpeciesFileGroup/taxonworks/issues/3989